Add FlexParams
This adds the ability to customize the behaviour of flex items.
It's actually not as big a change as I imagined; the main thing
is that it seprates out add_child from add_flex_child, and adds
a FlexParams type that is used in the second case; but since there
is a reasonable default, you can continue to pass a float to
methods that expect FlexParams, and it will work.

This does change default behaviour; children now have loose
constraints, and must use expand  explicitly if they want to be
forced to fill them.

This also includes a big overhaul of flex-related docs.

progress on #607